feat: add sort option to `list` command (#25)
This commit adds a `--sort-by` option to the `list` command. This option
allows the user to specify the fields by which to sort the list of
tickets. Multiple fields can be specified, separated by commas. The sort
order is the reverse of the order in which the fields are specified.
This makes for a more natural sort result when using multiple sort
options, e.g. `--sort-by namespace,name` will sort by name first, then
by namespace - this will result in a list that is sorted by namespace,
then by name.

The following fields can be specified:

* name
* namespace
* maprCluster
* maprUser
* creationTimestamp
* expiryTime

The default sort order is `namespace,name`.

Closes: #19