fix: handle duration overflow (> 292 years) (#26)
This commit fixes a bug where a duration that is greater than 292 years
would return a `<invalid>` string instead of a human readable string.
This is because the `time.Duration` type is an `int64` and uses nanoseconds
as the unit of time. The maximum value of an `int64` is 9223372036854775807
which is roughly 292 years. If the duration is greater than that, then
the `time.Duration` type be capped at the maximum value and the
`HumanDuration` function will return `<invalid>`.

This commit fixes this by checking if the duration is the maximum value
of an `int64` and if so, returning `inf` as the duration.

Closes: #23