chore: add helper script to enable kubectl completion
This is a workaround for enabling shell completion if the tool is used
as a kubectl plugin. Kubectl will automatically pick the script up if it
is named `kubectl_complete-<plugin-name>` and placed in some directory
that is in the `PATH` environment variable.

To enable completion for the `mapr-ticket` plugin, the user still has to
run `kubectl mapr-ticket completion <shell>` and source the output of
that command in some way.